Adicionando requisição da imagem

+++ app/Http/Controllers/PostController.php
@@ -89,7 +89,7 @@ class PostController extends Controller
 
              $data = $request->all();
 
-            if ($request->image->isValid()) {
+            if ($request->image && $request->image->isValid()) {
                 if (Storage::exists($post->image))
                     Storage::delete($post->image);
